requirements.txt
tensorflow-cpu==2.10.0
numpy<2.0
scikit-learn
mahjong
